requirements.txt
opencv-python
easygui
numpy
PyAutoGUI
pynput
pyperclip
pytesseract
pillow
pillow-heif